Adapt GitHub config for package upgrade workflow (#1)
Adapts the repository's GitHub config for a minimal package upgrade
workflow. The CODEOWNERS file has been removed, and any GitHub Actions
workflows have been removed that interacted with external services or
otherwise required upstream permissions. Other checks have been left in
place, since they are harmless and may be useful if we need to tweak
other things in the repo.

One new workflow has been added that runs tests for core packages, i.e.
all of the Atmosphere packages located in the `packages` directory. In
the upstream repo, this seems to be checked using Travis CI, but the
actual script for running tests headlessly lives in
`packages/test-in-console` which the new GHA workflow leverages.

The new workflow is configured to run on PRs and pushes, so we can run
the tests on PRs for upgrading NPM dependencies in various packages, and
so we can run the tests after merging into the default
`tulip-release-2.16` branch.