babel-compiler: Update json5 and @meteorjs/babel dependencies (#8)
There are a couple of change here. We first update dependencies of the
@meteorjs/babel npm package to resolve vulnerabilities. Some tests also
had to be fixed. See this
[commit](https://github.com/meteor/meteor/commit/18e5053b6cdb52cc329f97be6e74691d903edd41)
for the equivalent fixes in the original Meteor repo. Then, we make a
tarball of this package and upload it to the release
[here](https://github.com/tulip/meteor/releases/tag/updated-meteor-babel).
Finally, we update the babel-compiler package to use the tarball, along
with updating its json5 dependency.